requirements.txt
# Web Framework
fastapi==0.68.1
uvicorn==0.15.0
gunicorn==20.1.0
python-multipart==0.0.6
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
SQLAlchemy==1.4.23
alembic==1.7.1

# Face Recognition & Computer Vision
dlib==19.24.0
face-recognition==1.3.0
deepface==0.0.79
numpy==1.21.0
Pillow==8.3.2

# Utils
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-json-logger==2.0.2
requests==2.26.0

# Development
black==21.7b0
flake8==3.9.2

# Testing
pytest==6.2.5
pytest-cov==2.12.1
pytest-asyncio==0.15.1
httpx==0.19.0

# Production
supervisor==4.2.2

# Added from the code block
bcrypt==4.0.1 